+++ src/main/kotlin/shop/itbug/fluttercheckversionx/dsl/SettingPanel.kt
@@ -3,6 +3,7 @@ package shop.itbug.fluttercheckversionx.dsl
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.ui.DialogPanel
 import com.intellij.openapi.util.Disposer
+import com.intellij.ui.components.JBTextField
 import com.intellij.ui.dsl.builder.*
 import com.intellij.util.Alarm
 import shop.itbug.fluttercheckversionx.config.DoxListeningSetting
@@ -24,8 +25,9 @@ fun settingPanel(
 
     val alarm = Alarm(parentDisposable)
     val languageList = listOf("System", "中文", "繁體", "English", "한국어", "日本語")
-
+    lateinit var urlInput: Cell<JBTextField>
     val myPanel: DialogPanel = panel {
+
 
         buttonsGroup(PluginBundle.get("basic")) {
             row(PluginBundle.get("setting.language")) {
@@ -35,9 +37,24 @@ fun settingPanel(
         }.bind({ model.lang }, { model.lang = it })
 
         group(PluginBundle.get("check_flutter_version_title")) {
+
             row {
                 checkBox(PluginBundle.get("open")).bindSelected(dioxSetting::checkFlutterVersion)
             }.comment(PluginBundle.get("check_flutter_version_comment"))
+            row("Url") {
+                urlInput = textField().bindText(dioxSetting::checkFlutterVersionUrl).align(Align.FILL)
+            }
+            row {
+                button(PluginBundle.get("use_default_value")) {
+                    urlInput.component.text = Links.defaultFlutterInfosUrl
+                }
+                button("设置中国区镜像地址") {
+                    urlInput.component.text = Links.defaultFlutterInfosUrlByCN
+                }
+            }.comment(Links.generateDocCommit(urlInput.component.text, "View"))
+            row("pub.dev host") {
+                textField().bindText(dioxSetting::pubServerUrl).align(Align.FILL)
+            }
             documentCommentRow(Links.checkFlutterVersion)
         }