+++ src/main/kotlin/shop/itbug/fluttercheckversionx/services/FlutterService.kt
@@ -4,13 +4,14 @@ package shop.itbug.fluttercheckversionx.services
 import com.google.gson.Gson
 import com.google.gson.annotations.SerializedName
 import com.intellij.util.io.HttpRequests
+import shop.itbug.fluttercheckversionx.config.DioListingUiConfig
 
 class FlutterVersionCheckException(message: String) : Exception(message)
 
 object FlutterService {
     fun getVersion(): FlutterVersions {
         try {
-            val url = "https://storage.googleapis.com/flutter_infra_release/releases/releases_macos.json"
+            val url = DioListingUiConfig.setting.checkFlutterVersionUrl
             val get: String = HttpRequests.request(url).readString()
             return Gson().fromJson(get, FlutterVersions::class.java)
         } catch (e: Exception) {